# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/brands/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the brand profile for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brands"
                ],
                "summary": "Get Brand Profile",
                "responses": {
                    "200": {
                        "description": "profile details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create or update the brand profile for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "brands"
                ],
                "summary": "Update Brand Profile",
                "parameters": [
                    {
                        "description": "Brand Profile Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateBrandProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success and profile",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/jobs/feed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of open campaigns (job feed)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get Job Feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by niche",
                        "name": "niche",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of campaigns",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/jobs/my-applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of proposals submitted by the authenticated creator",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get My Applications",
                "responses": {
                    "200": {
                        "description": "List of proposals",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Proposal"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a proposal to a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Apply to a Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Application Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.applyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "proposal_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/upload/presigned": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a mock presigned URL for file uploads",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Get Presigned URL",
                "responses": {
                    "200": {
                        "description": "upload_url and public_url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Campaign": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "string"
                },
                "budget": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "platform": {
                    "description": "e.g. \"instagram\", \"youtube\"",
                    "type": "string"
                },
                "requirements": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "$ref": "#/definitions/domain.CampaignStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.CampaignStatus": {
            "type": "string",
            "enum": [
                "OPEN",
                "CLOSED"
            ],
            "x-enum-varnames": [
                "CampaignStatusOpen",
                "CampaignStatusClosed"
            ]
        },
        "domain.Proposal": {
            "type": "object",
            "properties": {
                "bid_amount": {
                    "type": "number"
                },
                "campaign": {
                    "$ref": "#/definitions/domain.Campaign"
                },
                "campaign_id": {
                    "type": "string"
                },
                "cover_note": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "proof_url": {
                    "type": "string"
                },
                "razorpay_order_id": {
                    "type": "string"
                },
                "razorpay_payment_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ProposalStatus"
                },
                "trial_video_url": {
                    "description": "or just video_url",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ProposalStatus": {
            "type": "string",
            "enum": [
                "APPLIED",
                "APPROVED",
                "FUNDED",
                "SUBMITTED",
                "COMPLETED",
                "PAID"
            ],
            "x-enum-varnames": [
                "ProposalStatusApplied",
                "ProposalStatusApproved",
                "ProposalStatusFunded",
                "ProposalStatusSubmitted",
                "ProposalStatusCompleted",
                "ProposalStatusPaid"
            ]
        },
...
        "http.applyRequest": {
            "type": "object",
            "properties": {
                "bid_amount": {
                    "type": "number"
                },
                "cover_letter": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "http.updateBrandProfileRequest": {
            "type": "object",
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "contact_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_in_company": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        }
    }
}